+++ turbofuro_runtime/src/actions/mod.rs
@@ -12,6 +12,7 @@ pub mod actors;
 pub mod alarms;
 pub mod convert;
 pub mod crypto;
+pub mod debug;
 pub mod form_data;
 pub mod fs;
 pub mod http_client;