+++ turbofuro_runtime/src/lib.rs
@@ -7,3 +7,6 @@ pub mod debug;
 pub mod evaluations;
 pub mod http_utils;
 pub mod resources;
+
+pub use crate::actions::kv::clean_kv;
+pub use crate::actions::kv::spawn_kv_cleaner;