+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 0.1.18 (pending)
+* Debug actions
+* Concurrent debug runs
+* KV expiration and increment support
+* Add get current date and time action
+* Fixed PubSub subscription being cancelled when the actor lags behind
 
 ## 0.1.17
 * Added support for HTTP cookies